+++ internal/domain/entity/vineyard.go
@@ -4,15 +4,15 @@ type Vineyard struct {
 	baseEntity
 }
 
-func NewVineyard(id int, name string) (*Vineyard, error) {
+func NewVineyard(id int64, name string) (*Vineyard, error) {
 	return &Vineyard{baseEntity{id: id, name: name}}, nil
 }
 
-func UnmarshalVineyardFromDatabase(id int, name string) (*Vineyard, error) {
+func UnmarshalVineyardFromDatabase(id int64, name string) (*Vineyard, error) {
 	return NewVineyard(id, name)
 }
 
-func (v Vineyard) ID() int {
+func (v Vineyard) ID() int64 {
 	return v.id
 }
